+++ src/components/button/index.js
@@ -1,7 +1,6 @@
 import Button from './Button';
 import ButtonGroup from './ButtonGroup';
-import IconButton from './IconButton';
 import SplitButton from './SplitButton';
 
 export default Button;
-export { Button, ButtonGroup, IconButton, SplitButton };
+export { Button, ButtonGroup, SplitButton };