+++ src/components/datagrid/datagrid.stories.js
@@ -1,22 +1,9 @@
 import React from 'react';
 import { addStoryInGroup, COMPOSITIONS } from '../../../.storybook/utils';
-import { boolean, number, select } from '@storybook/addon-knobs/react';
-import {
-  DataGrid,
-  Heading4,
-  IconMenu,
-  MenuItem,
-  Link,
-  StatusBullet,
-  TextSmall,
-  Tooltip,
-  Button,
-  ButtonGroup,
-} from '../../index';
+import { boolean, number } from '@storybook/addon-knobs/react';
+import { DataGrid, Heading4, IconMenu, MenuItem, Link, TextSmall, Button, ButtonGroup } from '../../index';
 
 import { rows1, rows2 } from '../../static/data/datagrid';
-
-const TooltippedStatusBullet = Tooltip(StatusBullet);
 
 const handleRowSelectionChange = (selectedRows) => {
   console.log(`onSelectionChange - selected row indexes: ${selectedRows}`);
@@ -42,7 +29,6 @@ export const basic = () => (
     selectable={boolean('Selectable', true)}
     comparableId={1}
     onSelectionChange={handleRowSelectionChange}
-    checkboxSize={select('Checkbox size', ['small', 'medium', 'large'], 'small')}
     processing={boolean('Processing', false)}
   >
     <DataGrid.HeaderRowOverlay
@@ -57,7 +43,6 @@ export const basic = () => (
     </DataGrid.HeaderRowOverlay>
 
     <DataGrid.HeaderRow>
-      <DataGrid.HeaderCell flex="min-width" />
       <DataGrid.HeaderCell onClick={() => console.log('onClick: column sort')} sorted="asc">
         Invoice
       </DataGrid.HeaderCell>
@@ -73,15 +58,6 @@ export const basic = () => (
     {rows1.map((row, index) => {
       return (
         <DataGrid.BodyRow key={index} onClick={(event) => console.log('onClick:', row.column5, event)}>
-          <DataGrid.Cell align="center" flex="min-width">
-            <TooltippedStatusBullet
-              color={row.column1}
-              tooltip={<TextSmall>Overdue</TextSmall>}
-              tooltipColor={row.column1}
-              tooltipSize="small"
-              size="medium"
-            />
-          </DataGrid.Cell>
           <DataGrid.Cell>
             <Link href="#" onClick={(event) => event.stopPropagation()} inherit={false}>
               {row.column5}
@@ -147,31 +123,23 @@ export const withFooter = () => (
     </DataGrid.HeaderRowOverlay>
 
     <DataGrid.HeaderRow>
-      <DataGrid.HeaderCell flex="min-width" />
-      <DataGrid.HeaderCell onClick={() => console.log('onClick: column sort')} sorted="asc">
+      <DataGrid.HeaderCell onClick={() => console.log('onClick: column sort')} sortable sorted="asc">
         Invoice
       </DataGrid.HeaderCell>
-      <DataGrid.HeaderCell onClick={() => console.log('onClick: column sort')} align="right">
+      <DataGrid.HeaderCell onClick={() => console.log('onClick: column sort')} sortable align="right">
         Amount
       </DataGrid.HeaderCell>
-      <DataGrid.HeaderCell flex="2" onClick={() => console.log('onClick: column sort')}>
+      <DataGrid.HeaderCell flex="2" onClick={() => console.log('onClick: column sort')} sortable>
         Customer
       </DataGrid.HeaderCell>
-      <DataGrid.HeaderCell onClick={() => console.log('onClick: column sort')}>Due date</DataGrid.HeaderCell>
+      <DataGrid.HeaderCell onClick={() => console.log('onClick: column sort')} sortable>
+        Due date
+      </DataGrid.HeaderCell>
       <DataGrid.HeaderCell flex="min-width" />
     </DataGrid.HeaderRow>
     {rows2.map((row, index) => {
       return (
         <DataGrid.BodyRow key={index} onClick={(event) => console.log('onClick:', row.column5, event)}>
-          <DataGrid.Cell align="center" flex="min-width">
-            <TooltippedStatusBullet
-              color={row.column1}
-              tooltip={<TextSmall>Overdue</TextSmall>}
-              tooltipColor={row.column1}
-              tooltipSize="small"
-              size="medium"
-            />
-          </DataGrid.Cell>
           <DataGrid.Cell>
             <Link href="#" onClick={(event) => event.stopPropagation()} inherit={false}>
               {row.column5}
@@ -223,31 +191,23 @@ export const withStickyColumns = () => (
     </DataGrid.HeaderRowOverlay>
 
     <DataGrid.HeaderRow>
-      <DataGrid.HeaderCell flex="min-width" />
-      <DataGrid.HeaderCell onClick={() => console.log('onClick: column sort')} sorted="asc">
+      <DataGrid.HeaderCell onClick={() => console.log('onClick: column sort')} sorted="asc" sortable>
         Invoice
       </DataGrid.HeaderCell>
-      <DataGrid.HeaderCell onClick={() => console.log('onClick: column sort')} align="right">
+      <DataGrid.HeaderCell onClick={() => console.log('onClick: column sort')} align="right" sortable>
         Amount
       </DataGrid.HeaderCell>
-      <DataGrid.HeaderCell flex="2" onClick={() => console.log('onClick: column sort')}>
+      <DataGrid.HeaderCell flex="2" onClick={() => console.log('onClick: column sort')} sortable>
         Customer
       </DataGrid.HeaderCell>
-      <DataGrid.HeaderCell onClick={() => console.log('onClick: column sort')}>Due date</DataGrid.HeaderCell>
+      <DataGrid.HeaderCell onClick={() => console.log('onClick: column sort')} sortable>
+        Due date
+      </DataGrid.HeaderCell>
       <DataGrid.HeaderCell flex="min-width" />
     </DataGrid.HeaderRow>
     {rows1.map((row, index) => {
       return (
         <DataGrid.BodyRow key={index} onClick={(event) => console.log('onClick:', row.column5, event)}>
-          <DataGrid.Cell align="center" flex="min-width">
-            <TooltippedStatusBullet
-              color={row.column1}
-              tooltip={<TextSmall>Overdue</TextSmall>}
-              tooltipColor={row.column1}
-              tooltipSize="small"
-              size="medium"
-            />
-          </DataGrid.Cell>
           <DataGrid.Cell>
             <Link href="#" onClick={(event) => event.stopPropagation()} inherit={false}>
               {row.column5}
@@ -296,31 +256,23 @@ export const advanced = () => (
     </DataGrid.HeaderRowOverlay>
 
     <DataGrid.HeaderRow>
-      <DataGrid.HeaderCell flex="min-width" />
-      <DataGrid.HeaderCell onClick={() => console.log('onClick: column sort')} sorted="asc">
+      <DataGrid.HeaderCell onClick={() => console.log('onClick: column sort')} sorted="asc" sortable>
         Invoice
       </DataGrid.HeaderCell>
-      <DataGrid.HeaderCell onClick={() => console.log('onClick: column sort')} align="right">
+      <DataGrid.HeaderCell onClick={() => console.log('onClick: column sort')} align="right" sortable>
         Amount
       </DataGrid.HeaderCell>
-      <DataGrid.HeaderCell flex="2" onClick={() => console.log('onClick: column sort')}>
+      <DataGrid.HeaderCell flex="2" onClick={() => console.log('onClick: column sort')} sortable>
         Customer
       </DataGrid.HeaderCell>
-      <DataGrid.HeaderCell onClick={() => console.log('onClick: column sort')}>Due date</DataGrid.HeaderCell>
+      <DataGrid.HeaderCell onClick={() => console.log('onClick: column sort')} sortable>
+        Due date
+      </DataGrid.HeaderCell>
       <DataGrid.HeaderCell flex="min-width" />
     </DataGrid.HeaderRow>
     {rows1.map((row, index) => {
       return (
         <DataGrid.BodyRow key={index} onClick={(event) => console.log('onClick:', row.column5, event)}>
-          <DataGrid.Cell align="center" flex="min-width">
-            <TooltippedStatusBullet
-              color={row.column1}
-              tooltip={<TextSmall>Overdue</TextSmall>}
-              tooltipColor={row.column1}
-              tooltipSize="small"
-              size="medium"
-            />
-          </DataGrid.Cell>
           <DataGrid.Cell>
             <Link href="#" onClick={(event) => event.stopPropagation()} inherit={false}>
               {row.column5}