+++ src/index.js
@@ -3,7 +3,7 @@ import Avatar, { AvatarStack } from './components/avatar';
 import Badge from './components/badge';
 import Banner from './components/banner';
 import Box from './components/box';
-import Button, { ButtonGroup, IconButton, SplitButton } from './components/button';
+import Button, { ButtonGroup, SplitButton } from './components/button';
 import {
   Heading1,
   Heading2,
@@ -31,6 +31,7 @@ import DetailPage, { DetailPageBody, DetailPageHeader } from './components/detai
 import Dialog, { DialogBase } from './components/dialog';
 import EmptyState from './components/emptyState';
 import Icon from './components/icon';
+import IconButton from './components/iconButton';
 import Input, { InputBase, NumericInput, Textarea, TimeInput } from './components/input';
 import Menu, { IconMenu, MenuItem, MenuDivider, MenuTitle } from './components/menu';
 import Link from './components/link';
@@ -52,6 +53,7 @@ import StatusBullet from './components/statusBullet';
 import StatusLabel from './components/statusLabel';
 import { IconTab, TitleTab, TabGroup } from './components/tab';
 import Tag from './components/tag';
+import Timer from './components/timer';
 import TimerPulser from './components/timerPulser';
 import { Toast, ToastContainer } from './components/toast';
 import Toggle from './components/toggle';
@@ -61,6 +63,7 @@ import Tooltip from './components/tooltip';
 import QTip from './components/qTip';
 import ProgressTracker from './components/progressTracker';
 import Widget from './components/widget';
+import WysiwygEditor from './components/wysiwygEditor';
 
 import {
   COLOR,
@@ -156,6 +159,7 @@ export {
   Tag,
   Textarea,
   TimeInput,
+  Timer,
   TimerPulser,
   TitleTab,
   Toast,
@@ -167,6 +171,7 @@ export {
   UITextSmall,
   ValidationText,
   Widget,
+  WysiwygEditor,
   QTip,
   COLOR,
   COLORS,